Fix: support 64-bit events discarded counter types
Previously, counters coming for 64-bit architectures would be treated in
the same way as 32-bit counters, and would therefore miss 32-bit
overflows (only happens if 2^32 events or more are lost within in a
packet).

+++ include/babeltrace/ctf-ir/metadata.h
@@ -52,9 +52,9 @@ struct ctf_stream_definition {
 	struct ctf_clock *current_clock;
 
 	/* Event discarded information */
-	uint32_t events_discarded;
-	uint64_t prev_timestamp;	/* Last event */
-	uint64_t prev_timestamp_end;	/* End-of-packet timestamp */
+	uint64_t events_discarded;
+	uint64_t prev_timestamp;	/* Start-of-last-packet timestamp */
+	uint64_t prev_timestamp_end;	/* End-of-last-packet timestamp */
 };
 
 struct ctf_event_definition {
